tests: fix version check for documentation
With the change to meson, the version string gets distributed from one
location (meson.build) and imported in all other locations, including
"doc/conf.py". Instead of checking the concrete value against the
other occurences, the new test checks if version is assigned from
"mesonconf.version" as a sanity check against regressions. Whenever
the mechanism for version consistency is changed, this test needs to
be updated as well.